backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
Flask-Mail==0.9.1
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.1.2
# face-recognition==1.3.0
# dlib==19.24.2

pandas==2.1.4
openpyxl==3.1.2
reportlab==4.0.9
requests==2.31.0
